move method to trait
